refactor: Rename query hooks
