+++ packages/sdk-react/src/components/expenses/ExpensesTable/UserTransactionsTable.tsx
@@ -1,6 +1,6 @@
 import { ReceiptPreview } from '../ReceiptPreview';
 import { TransactionDetails } from '../TransactionDetails';
-import { useGetTransactions } from '../hooks/useTransactions';
+import { useGetTransactions } from '../hooks/useGetTransactions';
 import { FILTER_TYPE_SEARCH, FILTER_TYPE_STARTED_AT } from './consts';
 import type { FilterTypes } from './types';
 import { type components } from '@/api';
